Refractered methods out of DAO and into a DBReaderService, to make the DAO even more generic
